moved shopify api to lib folder
