+++ lib/shopify.js
@@ -1,6 +1,3 @@
-// require("dotenv").config();
-// console.log("dotenv loaded");
-// console.log(process.env);
 export async function shopifyFetch({ query }) {
   const endpoint = process.env.SHOPIFY_STORE_DOMAIN;
   const key = process.env.SHOPIFY_STOREFRONT_ACCESS_TOKEN;
